pyproject.toml
[tool.poetry]
name = "fala-facil-api"
version = "0.1.0"
description = "API REST para o aplicativo Fala Fácil, uma plataforma de aprendizado de idiomas baseada em flashcards"
authors = ["Elvis Mariano <admelvismariano@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = "^5.0.1"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
dj-database-url = "^2.1.0"
drf-yasg = "^1.21.7"
django-filter = "^24.1"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"
coreapi = "^2.3.3"
django-debug-toolbar = "^4.3.0"
django-health-check = "^3.18.3"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src --cov-report=term-missing"
testpaths = ["tests"]
